requirements.txt
huggingface-hub==0.29.1
transformers==4.45.1
numpy==2.2.3
hydra-core==1.3
hydra_colorlog==1.2.0
torch==2.4.1
datasets==3.0.1
accelerate==0.34.2
bitsandbytes==0.44.1
rouge-score==0.1.2
scipy==1.14.1
tensorboard==2.18.0
scikit-learn==1.5.2
deepspeed==0.15.4
langchain-community
langchain
langchain_huggingface
sentence-transformers
faiss-cpu==1.13.2
